+++ math3d/transform.py
@@ -42,7 +42,7 @@ class Transform(object):
         # if len(arg) == 1 and isSequence(arg[0]):
         #     self.__createOnSequence(arg[0])
         if type(arg) in (tuple, list):
-            self.__create_on_sequence(np.array(arg, dtype=np.float32))
+            self.__create_on_sequence(np.array(arg, dtype=np.float64))
         elif type(arg) == np.ndarray and arg.shape in ((4,4), (3,4)):
             self._o = m3d.Orientation(arg[:3,:3])
             self._v = m3d.Vector(arg[:3,3])
@@ -84,7 +84,7 @@ class Transform(object):
             self._v = m3d.Vector(args[3])
         elif len(args) == 12:
             # // 12 numbers are required
-            args = np.array(args, dtype=float32)
+            args = np.array(args, dtype=float64)
             self._o = m3d.Orientation(args[:9])
             self._v = m3d.Vector(args[9:])
         else:
